Split various payload definition
It is starting to be hard to work on those payload so
splitting them to make my life easier without changing the SDK's API
