+++ src/cnd/cnd.ts
@@ -3,7 +3,15 @@ import actionToHttpRequest, {
   FieldValueResolverFn
 } from "./action_to_http_request";
 import { problemResponseInterceptor } from "./axios_rfc7807_middleware";
-import { Action, EmbeddedRepresentationSubEntity, Entity } from "./siren";
+import { SwapSubEntity } from "./rfc003_payload";
+import { Action, Entity } from "./siren";
+import {
+  HalightLightningBitcoinHanEthereumEtherRequestBody,
+  HalightLightningBitcoinHerc20EthereumErc20RequestBody,
+  HanEthereumEtherHalightLightningBitcoinRequestBody,
+  Herc20EthereumErc20HalightLightningBitcoinRequestBody,
+  SwapRequest
+} from "./swaps_payload";
 
 interface GetInfo {
   id: string;
@@ -29,304 +37,6 @@ export interface Asset {
 export interface Peer {
   peer_id: string;
   address_hint?: string;
-}
-
-/**
- * The parameters that the two parties have agreed on for swap execution.
- *
- * The swap request is sent by the party in the role of Alice to initiate a swap with Bob.
- */
-export interface SwapRequest {
-  alpha_ledger: Ledger;
-  alpha_asset: Asset;
-  beta_ledger: Ledger;
-  beta_asset: Asset;
-  alpha_expiry?: number;
-  beta_expiry?: number;
-  alpha_ledger_refund_identity?: string;
-  beta_ledger_redeem_identity?: string;
-  peer: Peer;
-}
-
-interface CoreRequestBody<A, B> {
-  alpha: A;
-  beta: B;
-  role: "Alice" | "Bob";
-  peer: Peer;
-}
-
-export interface RequestParams {
-  amount: string;
-  identity: string;
-}
-
-interface Han {
-  absolute_expiry: number;
-}
-
-interface Herc20 {
-  contract_address: string;
-  absolute_expiry: number;
-}
-
-interface Halight {
-  cltv_expiry: number;
-}
-
-interface Bitcoin {
-  network: string;
-}
-
-interface Ethereum {
-  chain_id: number;
-}
-
-export type HanEthereumEtherRequestParams = RequestParams & Han & Ethereum;
-export type HalightLightningBitcoinRequestParams = RequestParams &
-  Halight &
-  Bitcoin;
-export type Herc20EthereumErc20RequestParams = RequestParams &
-  Herc20 &
-  Ethereum;
-
-export type HanEthereumEtherHalightLightningBitcoinRequestBody = CoreRequestBody<
-  HanEthereumEtherRequestParams,
-  HalightLightningBitcoinRequestParams
->;
-export type Herc20EthereumErc20HalightLightningBitcoinRequestBody = CoreRequestBody<
-  Herc20EthereumErc20RequestParams,
-  HalightLightningBitcoinRequestParams
->;
-export type HalightLightningBitcoinHanEthereumEtherRequestBody = CoreRequestBody<
-  HalightLightningBitcoinRequestParams,
-  HanEthereumEtherRequestParams
->;
-export type HalightLightningBitcoinHerc20EthereumErc20RequestBody = CoreRequestBody<
-  HalightLightningBitcoinRequestParams,
-  Herc20EthereumErc20RequestParams
->;
-
-export interface BitcoinSendAmountToAddressPayload {
-  to: string;
-  amount: string;
-  network: string;
-}
-
-export interface BitcoinBroadcastSignedTransactionPayload {
-  hex: string;
-  network: string;
-}
-
-export interface EthereumDeployContractPayload {
-  data: string;
-  amount: string;
-  gas_limit: string;
-  network: string;
-}
-
-export interface EthereumCallContractPayload {
-  contract_address: string;
-  data: string;
-  gas_limit: string;
-  network: string;
-}
-
-export interface LndSendPaymentPayload {
-  self_public_key: string;
-  to_public_key: string;
-  amount: string;
-  secret_hash: string;
-  final_cltv_delta: number;
-  chain: string;
-  network: string;
-}
-
-export interface LndAddHoldInvoicePayload {
-  self_public_key: string;
-  amount: string;
-  secret_hash: string;
-  expiry: number;
-  cltv_expiry: number;
-  chain: string;
-  network: string;
-}
-
-export interface LndSettleInvoicePayload {
-  self_public_key: string;
-  secret: string;
-  chain: string;
-  network: string;
-}
-
-export type LedgerAction =
-  | {
-      type: "bitcoin-send-amount-to-address";
-      payload: BitcoinSendAmountToAddressPayload;
-    }
-  | {
-      type: "bitcoin-broadcast-signed-transaction";
-      payload: BitcoinBroadcastSignedTransactionPayload;
-    }
-  | {
-      type: "ethereum-deploy-contract";
-      payload: EthereumDeployContractPayload;
-    }
-  | {
-      type: "ethereum-call-contract";
-      payload: EthereumCallContractPayload;
-    }
-  | {
-      type: "lnd-send-payment";
-      payload: LndSendPaymentPayload;
-    }
-  | {
-      type: "lnd-add-hold-invoice";
-      payload: LndAddHoldInvoicePayload;
-    }
-  | {
-      type: "lnd-settle-invoice";
-      payload: LndSettleInvoicePayload;
-    }
-  | {
-      type: string;
-      payload: any;
-    };
-
-export interface SwapDetails extends Entity {
-  properties?: SwapProperties;
-}
-
-export interface SwapSubEntity extends EmbeddedRepresentationSubEntity {
-  properties?: SwapProperties;
-}
-
-export interface SwapProperties {
-  /**
-   * The id of the swap.
-   */
-  id: string;
-  /**
-   * The peer-id of the counterparty of this swap.
-   */
-  counterparty: string;
-  /**
-   * The role in which you are participating in this swap.
-   */
-  role: "Alice" | "Bob";
-  /**
-   * The cryptographic protocol that is employed in this swap.
-   */
-  protocol: string;
-  /**
-   * The status this swap is currently in.
-   */
-  status: "IN_PROGRESS" | "SWAPPED" | "NOT_SWAPPED" | "INTERNAL_FAILURE";
-  /**
-   * The parameters of this swap.
-   */
-  parameters: {
-    alpha_asset: Asset;
-    alpha_ledger: Ledger;
-    beta_asset: Asset;
-    beta_ledger: Ledger;
-    [k: string]: any;
-  };
-  /**
-   * The detailed state of the swap.
-   */
-  state?: {
-    /**
-     * The state of the alpha ledgers regarding the swap.
-     */
-    alpha_ledger: {
-      /**
-       * The transaction ID of the deployment transaction on the alpha ledger.
-       */
-      deploy_tx: string | null;
-      /**
-       * The transaction ID of the funding transaction on the alpha ledger.
-       */
-      fund_tx: string | null;
-      /**
-       * The address of the htlc on the alpha ledger.
-       */
-      htlc_location: any;
-      /**
-       * The transaction ID of the redeem transaction on the alpha ledger.
-       */
-      redeem_tx: string | null;
-      /**
-       * The transaction ID of the refund transaction on the alpha ledger.
-       */
-      refund_tx: string | null;
-      /**
-       * The status of the HTLC on the alpha ledgers.
-       */
-      status:
-        | "NOT_DEPLOYED"
-        | "DEPLOYED"
-        | "FUNDED"
-        | "REDEEMED"
-        | "REFUNDED"
-        | "INCORRECTLY_FUNDED";
-      [k: string]: any;
-    };
-    /**
-     * The state of the beta ledgers regarding the swap.
-     */
-    beta_ledger: {
-      /**
-       * The transaction ID of the deployment transaction on the beta ledger.
-       */
-      deploy_tx: string | null;
-      /**
-       * The transaction ID of the funding transaction on the beta ledger.
-       */
-      fund_tx: string | null;
-      /**
-       * The address of the htlc on the beta ledger.
-       */
-      htlc_location: any;
-      /**
-       * The transaction ID of the redeem transaction on the beta ledger.
-       */
-      redeem_tx: string | null;
-      /**
-       * The transaction ID of the refund transaction on the beta ledger.
-       */
-      refund_tx: string | null;
-      /**
-       * The status of the HTLC on the beta ledgers.
-       */
-      status:
-        | "NOT_DEPLOYED"
-        | "DEPLOYED"
-        | "FUNDED"
-        | "REDEEMED"
-        | "REFUNDED"
-        | "INCORRECTLY_FUNDED";
-      [k: string]: any;
-    };
-    /**
-     * The state created during the communication of the two parties regarding the swap.
-     */
-    communication: {
-      /**
-       * The expiry value of the HTLC on the alpha ledgers. The semantic value depends on the ledgers.
-       */
-      alpha_expiry: number;
-      alpha_redeem_identity: string | null;
-      alpha_refund_identity: string | null;
-      beta_expiry: number;
-      beta_redeem_identity: string | null;
-      beta_refund_identity: string | null;
-      secret_hash: string;
-      status: "SENT" | "ACCEPTED" | "DECLINED";
-      [k: string]: any;
-    };
-    [k: string]: any;
-  };
-  [k: string]: any;
 }
 
 /**
